Rename types to models
