+++ streamlit_webrtc/mix.py
@@ -16,8 +16,8 @@ from aiortc.contrib.media import RelayStreamTrack
 from aiortc.mediastreams import MediaStreamError
 
 from .eventloop import get_server_event_loop, loop_context
+from .models import MixerBase, MixerT
 from .relay import get_global_relay
-from .types import MixerBase, MixerT
 
 __all__ = [
     "MixerBase",
